user permissions resolved

+++ microblog/views.py
@@ -74,19 +74,17 @@ def view_blog(request, blog_slug):
 
 @active_admin_required
 def blog_add(request):
-    form = BlogPostForm()
+    form = BlogPostForm(request.GET, is_superuser=request.user.is_superuser)
     tags_list = Tags.objects.all()
     categories_list = Category.objects.filter(is_active=True)
     if request.method == "POST":
-        form = BlogPostForm(request.POST)
+        form = BlogPostForm(request.POST, is_superuser=request.user.is_superuser)
         if form.is_valid():
             blog_post = form.save(commit=False)
             blog_post.user = request.user
-            blog_post.status = 'Drafted'
-            if request.POST.get('status') == 'Published':
-                blog_post.status = 'Published'
-            elif request.POST.get('status') == 'Rejected':
-                blog_post.status = 'Rejected'
+            #blog_post.status = 'Drafted'
+            if request.user.is_superuser:
+                blog_post.status = request.POST.get('status')
             blog_post.save()
 
             if request.POST.get('tags', ''):
@@ -111,46 +109,48 @@ def blog_add(request):
 @active_admin_required
 def edit_blog(request, blog_slug):
     blog_name = Post.objects.get(slug=blog_slug)
-    form = BlogPostForm(instance=blog_name)
+    if blog_name.user == request.user or request.user.is_superuser == True:
+        form = BlogPostForm(instance=blog_name)
 
-    categories_list = Category.objects.filter(is_active=True)
-    if request.method == "POST":
-        form = BlogPostForm(request.POST, instance=blog_name)
-        if form.is_valid():
-            blog_post = form.save(commit=False)
-            blog_post.user = request.user
-            blog_post.status = 'Drafted'
-            if request.POST.get('status') == 'Published':
-                blog_post.status = 'Published'
-            elif request.POST.get('status') == 'Rejected':
-                blog_post.status = 'Rejected'
-            blog_post.save()
+        categories_list = Category.objects.filter(is_active=True)
+        if request.method == "POST":
+            form = BlogPostForm(request.POST, instance=blog_name)
+            if form.is_valid():
+                blog_post = form.save(commit=False)
+                blog_post.user = request.user
+                blog_post.status = 'Drafted'
+                if request.POST.get('status') == 'Published':
+                    blog_post.status = 'Published'
+                elif request.POST.get('status') == 'Rejected':
+                    blog_post.status = 'Rejected'
+                blog_post.save()
 
-            if request.POST.get('tags', ''):
-                tags = request.POST.get('tags')
+                if request.POST.get('tags', ''):
+                    tags = request.POST.get('tags')
 
-                splitted = tags.split(',')
-                for s in splitted:
-                    final = s.strip()
-                    if not Tags.objects.filter(name=final).exists():
-                        Tags.objects.create(name=final)
+                    splitted = tags.split(',')
+                    for s in splitted:
+                        final = s.strip()
+                        if not Tags.objects.filter(name=final).exists():
+                            Tags.objects.create(name=final)
 
-            messages.success(request, 'Successfully updated your blog post')
-            data = {'error': False, 'response': 'Successfully updated your blog post'}
-        else:
-            data = {'error': True, 'response': form.errors}
-        return HttpResponse(json.dumps(data))
-    context = {'form': form, 'blog_name': blog_name, 'status_choices': STATUS_CHOICE,
-               'categories_list': categories_list}
-    return render(request, 'dashboard/blog/blog_add.html', context)
+                messages.success(request, 'Successfully updated your blog post')
+                data = {'error': False, 'response': 'Successfully updated your blog post'}
+            else:
+                data = {'error': True, 'response': form.errors}
+            return HttpResponse(json.dumps(data))
+        context = {'form': form, 'blog_name': blog_name, 'status_choices': STATUS_CHOICE,
+                   'categories_list': categories_list}
+        return render(request, 'dashboard/blog/blog_add.html', context)
 
 
 @active_admin_required
 def delete_blog(request, blog_slug):
     blog_name = Post.objects.get(slug=blog_slug)
-    blog_name.delete()
-    messages.success(request, 'Blog successfully deleted')
-    return HttpResponseRedirect('/dashboard/blog/')
+    if blog_name.user == request.user or request.user.is_superuser == True:
+        blog_name.delete()
+        messages.success(request, 'Blog successfully deleted')
+        return HttpResponseRedirect('/dashboard/blog/')
 
 
 @active_admin_required
@@ -196,71 +196,75 @@ def add_category(request):
 @active_admin_required
 def edit_category(request, category_slug):
     category_name = Category.objects.get(slug=category_slug)
-    form = BlogCategoryForm(instance=category_name)
+    if category_name.user == request.user or request.user.is_superuser == True:
+        form = BlogCategoryForm(instance=category_name)
 
-    if request.method == 'POST':
-        form = BlogCategoryForm(request.POST, instance=category_name)
-        if form.is_valid():
-            form.save()
-            messages.success(request, 'Successfully updated your category')
-            data = {'error': False, 'response': 'Successfully updated your category'}
-        else:
-            data = {'error': True, 'response': form.errors}
-        return HttpResponse(json.dumps(data))
-    context = {'form': form, 'category_name': category_name}
-    return render(request, 'dashboard/category/category_add.html', context)
+        if request.method == 'POST':
+            form = BlogCategoryForm(request.POST, instance=category_name)
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Successfully updated your category')
+                data = {'error': False, 'response': 'Successfully updated your category'}
+            else:
+                data = {'error': True, 'response': form.errors}
+            return HttpResponse(json.dumps(data))
+        context = {'form': form, 'category_name': category_name}
+        return render(request, 'dashboard/category/category_add.html', context)
 
 
 @active_admin_required
 def delete_category(request, category_slug):
     category = Category.objects.get(slug=category_slug)
-    category.delete()
-    return HttpResponseRedirect('/dashboard/category/')
+    if category.user == request.user or request.user.is_superuser == True:
+        category.delete()
+        return HttpResponseRedirect('/dashboard/category/')
 
 
 @active_admin_required
 def bulk_actions_blog(request):
-    if request.method == 'GET':
-        if 'blog_ids[]' in request.GET:
+    if request.user.is_superuser:
+        if request.method == 'GET':
+            if 'blog_ids[]' in request.GET:
 
-            if request.GET.get('action') == 'Published' or request.GET.get('action') == 'Drafted' or request.GET.get(
-                    'action') == 'Rejected':
-                Post.objects.filter(id__in=request.GET.getlist('blog_ids[]')).update(
-                    status=request.GET.get('action'))
-                messages.success(request,
-                                 'Selected blog posts successfully updated as ' + str(request.GET.get('action')))
+                if request.GET.get('action') == 'Published' or request.GET.get('action') == 'Drafted' or request.GET.get(
+                        'action') == 'Rejected':
+                    Post.objects.filter(id__in=request.GET.getlist('blog_ids[]')).update(
+                        status=request.GET.get('action'))
+                    messages.success(request,
+                                     'Selected blog posts successfully updated as ' + str(request.GET.get('action')))
 
-            elif request.GET.get('action') == 'Delete':
-                Post.objects.filter(id__in=request.GET.getlist('blog_ids[]')).delete()
+                elif request.GET.get('action') == 'Delete':
+                    Post.objects.filter(id__in=request.GET.getlist('blog_ids[]')).delete()
 
-            return HttpResponse(json.dumps({'response': True}))
-        else:
-            messages.warning(request, 'Please select at-least one record to perform this action')
-            return HttpResponse(json.dumps({'response': False}))
+                return HttpResponse(json.dumps({'response': True}))
+            else:
+                messages.warning(request, 'Please select at-least one record to perform this action')
+                return HttpResponse(json.dumps({'response': False}))
 
-    return render(request, '/dashboard/blog/')
+        return render(request, '/dashboard/blog/')
 
 
 @active_admin_required
 def bulk_actions_category(request):
-    if request.method == 'GET':
-        if 'blog_ids[]' in request.GET:
-            if request.GET.get('action') == 'True':
-                Category.objects.filter(id__in=request.GET.getlist('blog_ids[]')).update(
-                    is_active=True)
-                messages.success(request, 'Selected Categories successfully updated as Active')
-            elif request.GET.get('action') == 'False':
-                Category.objects.filter(id__in=request.GET.getlist('blog_ids[]')).update(
-                    is_active=False)
-                messages.success(request, 'Selected Categories successfully updated as Inactive')
+    if request.user.is_superuser:
+        if request.method == 'GET':
+            if 'blog_ids[]' in request.GET:
+                if request.GET.get('action') == 'True':
+                    Category.objects.filter(id__in=request.GET.getlist('blog_ids[]')).update(
+                        is_active=True)
+                    messages.success(request, 'Selected Categories successfully updated as Active')
+                elif request.GET.get('action') == 'False':
+                    Category.objects.filter(id__in=request.GET.getlist('blog_ids[]')).update(
+                        is_active=False)
+                    messages.success(request, 'Selected Categories successfully updated as Inactive')
 
-            elif request.GET.get('action') == 'Delete':
-                Category.objects.filter(id__in=request.GET.getlist('blog_ids[]')).delete()
-                messages.success(request, 'Selected Categories successfully deleted!')
+                elif request.GET.get('action') == 'Delete':
+                    Category.objects.filter(id__in=request.GET.getlist('blog_ids[]')).delete()
+                    messages.success(request, 'Selected Categories successfully deleted!')
 
-            return HttpResponse(json.dumps({'response': True}))
-        else:
-            messages.warning(request, 'Please select at-least one record to perform this action')
-            return HttpResponse(json.dumps({'response': False}))
+                return HttpResponse(json.dumps({'response': True}))
+            else:
+                messages.warning(request, 'Please select at-least one record to perform this action')
+                return HttpResponse(json.dumps({'response': False}))
 
-    return render(request, '/dashboard/category/')
+        return render(request, '/dashboard/category/')
